+++ examples/proxy.rs
@@ -1,6 +1,6 @@
 use std::error::Error;
 
-use ureq::{config::Config, Agent, Proxy};
+use ureq::{Agent, Proxy, config::Config};
 
 // Use this example with something like mitmproxy
 // $ mitmproxy --listen-port 8080