+++ src/body/charset.rs
@@ -138,8 +138,8 @@ mod test {
     #[test]
     #[cfg(feature = "charset")]
     fn non_ascii_reason() {
-        use crate::test::init_test_log;
         use crate::Agent;
+        use crate::test::init_test_log;
 
         init_test_log();
         let agent: Agent = Agent::config_builder().max_redirects(0).build().into();