+++ src/config.rs
@@ -6,8 +6,8 @@ use std::sync::Arc;
 use std::time::Duration;
 
 use crate::middleware::{Middleware, MiddlewareChain};
-use crate::{http, Body, Error};
 use crate::{Agent, AsSendBody, Proxy, RequestBuilder};
+use crate::{Body, Error, http};
 
 #[cfg(feature = "_tls")]
 use crate::tls::TlsConfig;
@@ -792,11 +792,7 @@ impl AutoHeaderValue {
             AutoHeaderValue::Provided(v) => v.as_str(),
         };
 
-        if x.is_empty() {
-            None
-        } else {
-            Some(x)
-        }
+        if x.is_empty() { None } else { Some(x) }
     }
 }
 
@@ -812,33 +808,33 @@ impl<S: AsRef<str>> From<S> for AutoHeaderValue {
 impl ConfigBuilder<AgentScope> {
     /// Finalize the config
     pub fn build(self) -> Config {
-        self.0 .0
+        self.0.0
     }
 }
 
 impl<Any> ConfigBuilder<RequestScope<Any>> {
     /// Finalize the config
     pub fn build(self) -> RequestBuilder<Any> {
-        self.0 .0
+        self.0.0
     }
 }
 
 impl<S: AsSendBody> ConfigBuilder<HttpCrateScope<S>> {
     /// Finalize the config
     pub fn build(self) -> http::Request<S> {
-        self.0 .0
+        self.0.0
     }
 }
 
 impl<'a, S: AsSendBody> ConfigBuilder<RequestExtScope<'a, S>> {
     /// Finalize the config
     pub fn build(self) -> WithAgent<'a, S> {
-        self.0 .0
+        self.0.0
     }
 
     /// Run the request with the agent in the ConfigBuilder
     pub fn run(self) -> Result<Response<Body>, Error> {
-        self.0 .0.run()
+        self.0.0.run()
     }
 }
 
@@ -856,7 +852,7 @@ pub struct Timeouts {
     /// Max duration for doing the DNS lookup when establishing the connection
     pub resolve: Option<Duration>,
 
-    /// Max duration for establishing the connection
+    /// Max duration for establishing the connection.
     pub connect: Option<Duration>,
 
     /// Max duration for sending the request, but not the request body.